+++ src/flags/bool.go
@@ -6,9 +6,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// boolFlag provides access to boolean Cobra command-line flags
-// in a way where Go's usage checker (which produces compilation errors for unused variables)
-// enforces that the programmer didn't forget to define or read the flag.
+// Bool provides mistake-safe access to boolean Cobra command-line flags.
 func Bool(name, short, desc string) (AddFunc, readBoolFlagFunc) {
 	addFlag := func(cmd *cobra.Command) {
 		cmd.PersistentFlags().BoolP(name, short, false, desc)
