+++ src/flags/core.go
@@ -1,4 +1,6 @@
-// Package flags provides helper methods for working with Cobra flags.
+// Package flags provides helper methods for working with Cobra flags
+// in a way where Go's usage checker (which produces compilation errors for unused variables)
+// enforces that the programmer didn't forget to define or read the flag.
 package flags
 
 import (
